+++ tests/Functional/ConnectionsTest.php
@@ -16,6 +16,7 @@ use AppBundle\Model\Category;
 use AppBundle\Model\Conn2Book;
 use AppBundle\Model\Isbn;
 use Illuminate\Database\Schema\Blueprint;
+use Symfony\Bridge\PhpUnit\SetUpTearDownTrait;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use WouterJ\EloquentBundle\Facade\Schema;
 
@@ -24,12 +25,14 @@ use WouterJ\EloquentBundle\Facade\Schema;
  */
 class ConnectionsTest extends KernelTestCase
 {
+    use SetUpTearDownTrait;
+
     protected static function getKernelClass()
     {
         return 'TestKernel';
     }
 
-    protected function setUp()
+    protected function doSetUp()
     {
         static::bootKernel();