+++ tests/Functional/RelationshipsTest.php
@@ -15,6 +15,7 @@ use AppBundle\Model\Book;
 use AppBundle\Model\Category;
 use AppBundle\Model\Isbn;
 use Illuminate\Database\Schema\Blueprint;
+use Symfony\Bridge\PhpUnit\SetUpTearDownTrait;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use WouterJ\EloquentBundle\Facade\Schema;
 
@@ -23,12 +24,14 @@ use WouterJ\EloquentBundle\Facade\Schema;
  */
 class RelationshipsTest extends KernelTestCase
 {
+    use SetUpTearDownTrait;
+
     protected static function getKernelClass()
     {
         return 'TestKernel';
     }
 
-    protected function setUp()
+    protected function doSetUp()
     {
         static::bootKernel();