+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+ * Dropped Symfony 2.8 support
+ * Dropped PHP 7.0 support
+
 1.0.2
 -----